+++ google/gcloud/cluster.go
@@ -1,0 +1,7 @@
+package gcloud
+
+type Cluster struct {
+	Project string `json:"project" yaml:"project"`
+	Region  string `json:"region" yaml:"region"`
+	Name    string `json:"name" yaml:"name"`
+}
