+++ google/gcloud/config.go
@@ -20,11 +20,6 @@ func (c Config) FindCluster(envName, clusterName string) (Cluster, bool) {
 	return Cluster{}, false
 }
 
-type Environment struct {
-	Name     string    `json:"name" yaml:"name"`
-	Clusters []Cluster `json:"clusters" yaml:"clusters"`
-}
-
 func (c Config) ClusterNames() []string {
 	var ret []string
 	for _, account := range c.Environments {
@@ -46,10 +41,4 @@ func (c Config) ClusterNamesForEnv(envName string) []string {
 		}
 	}
 	return nil
-}
-
-type Cluster struct {
-	Project string `json:"project" yaml:"project"`
-	Region  string `json:"region" yaml:"region"`
-	Name    string `json:"name" yaml:"name"`
 }
