+++ google/gcloud/environment.go
@@ -1,0 +1,6 @@
+package gcloud
+
+type Environment struct {
+	Name     string    `json:"name" yaml:"name"`
+	Clusters []Cluster `json:"clusters" yaml:"clusters"`
+}
